raspberry-pi/requirements.txt
evdev>=1.7.0
requests>=2.31.0
RPi.GPIO>=0.7.1
